Don't create a static library as an intermediate step
I've discovered how to compile the tinyrenderer code with the main code
without the intermediate step of building a static library for the
tinyrenderer code, and it's much simpler this way. Also solves the
problem of "undefined reference" errors when linking that I wasn't able
to solve when tinyrenderer was compiled to a static library.