Fix pyramid coloring bug caused by WebGL assuming pre-multiplied alpha
This commit fixes a bug with browsers assuming pre-multiplied alpha in
the back buffer (unlike OpenGL which does not). This caused the pyramid
to be oddly colored depending on the alpha values of its vertices. The
solution was to turn off alpha in the back buffer (see option #2 in this
list of fixes:
https://webglfundamentals.org/webgl/lessons/webgl-and-alpha.html).